[rename] signupTerms 폴더 signup으로 폴더이름 변경
