+++ src/components/signup/Terms.tsx
